skalowanie prędkości w zależności od wymiarów ekranu

+++ snakegame/Vector.java
@@ -41,7 +41,7 @@ public class Vector {
         y -= v.y;
     }
 
-    public void limit(float max) {
+    public void limit(double max) {
         if(length() > max) {
             normalize();
             multiply(max);
